requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
accelerate==0.18.0
    # via -r requirements.in
aiohappyeyeballs==2.4.4
    # via aiohttp
aiohttp==3.11.10
    # via
    #   datasets
    #   fsspec
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
async-timeout==5.0.1
    # via aiohttp
attrs==24.2.0
    # via
    #   aiohttp
    #   jsonlines
beartype==0.19.0
    # via st-moe-pytorch
blis==1.0.1
    # via thinc
catalogue==2.0.10
    # via
    #   spacy
    #   srsly
    #   thinc
certifi==2024.8.30
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via
    #   typer
    #   wandb
cloudpathlib==0.20.0
    # via weasel
colt5-attention==0.11.1
    # via st-moe-pytorch
confection==0.1.5
    # via
    #   thinc
    #   weasel
cymem==2.0.10
    # via
    #   preshed
    #   spacy
    #   thinc
datasets==3.1.0
    # via evaluate
dill==0.3.8
    # via
    #   datasets
    #   evaluate
    #   multiprocess
docker-pycreds==0.4.0
    # via wandb
einops==0.8.0
    # via
    #   -r requirements.in
    #   colt5-attention
    #   local-attention
    #   st-moe-pytorch
evaluate==0.4.3
    # via -r requirements.in
filelock==3.16.1
    # via
    #   datasets
    #   huggingface-hub
    #   torch
    #   transformers
    #   triton
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2024.9.0
    # via
    #   datasets
    #   evaluate
    #   huggingface-hub
    #   torch
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via wandb
huggingface-hub==0.26.3
    # via
    #   accelerate
    #   datasets
    #   evaluate
    #   miditok
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   requests
    #   yarl
jinja2==3.1.4
    # via
    #   spacy
    #   torch
jsonlines==4.0.0
    # via -r requirements.in
langcodes==3.5.0
    # via spacy
language-data==1.3.0
    # via langcodes
local-attention==1.9.15
    # via colt5-attention
marisa-trie==1.2.1
    # via language-data
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
miditok==3.0.3
    # via -r requirements.in
mpmath==1.3.0
    # via sympy
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.16
    # via
    #   datasets
    #   evaluate
murmurhash==1.0.11
    # via
    #   preshed
    #   spacy
    #   thinc
networkx==3.4.2
    # via torch
numpy==2.0.2
    # via
    #   -r requirements.in
    #   accelerate
    #   blis
    #   datasets
    #   evaluate
    #   miditok
    #   pandas
    #   spacy
    #   symusic
    #   thinc
    #   transformers
nvidia-cublas-cu12==12.4.5.8
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.4.127
    # via torch
nvidia-cuda-nvrtc-cu12==12.4.127
    # via torch
nvidia-cuda-runtime-cu12==12.4.127
    # via torch
nvidia-cudnn-cu12==9.1.0.70
    # via torch
nvidia-cufft-cu12==11.2.1.3
    # via torch
nvidia-curand-cu12==10.3.5.147
    # via torch
nvidia-cusolver-cu12==11.6.1.9
    # via torch
nvidia-cusparse-cu12==12.3.1.170
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-nccl-cu12==2.21.5
    # via torch
nvidia-nvjitlink-cu12==12.4.127
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.4.127
    # via torch
packaging==24.2
    # via
    #   accelerate
    #   colt5-attention
    #   datasets
    #   evaluate
    #   huggingface-hub
    #   spacy
    #   thinc
    #   transformers
    #   weasel
pandas==2.2.3
    # via
    #   datasets
    #   evaluate
platformdirs==4.3.6
    # via
    #   symusic
    #   wandb
preshed==3.0.9
    # via
    #   spacy
    #   thinc
propcache==0.2.1
    # via
    #   aiohttp
    #   yarl
protobuf==5.29.1
    # via wandb
psutil==6.1.0
    # via
    #   accelerate
    #   wandb
pyarrow==18.1.0
    # via datasets
pydantic==2.10.3
    # via
    #   confection
    #   spacy
    #   thinc
    #   wandb
    #   weasel
pydantic-core==2.27.1
    # via pydantic
pygments==2.18.0
    # via rich
pysmartdl==1.3.4
    # via symusic
python-dateutil==2.9.0.post0
    # via pandas
pytz==2024.2
    # via pandas
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   accelerate
    #   datasets
    #   huggingface-hub
    #   transformers
    #   wandb
regex==2024.11.6
    # via transformers
requests==2.32.3
    # via
    #   datasets
    #   evaluate
    #   huggingface-hub
    #   spacy
    #   transformers
    #   wandb
    #   weasel
rich==13.9.4
    # via typer
safetensors==0.4.5
    # via
    #   accelerate
    #   transformers
sentry-sdk==2.19.2
    # via wandb
sentencepiece==0.2.0

setproctitle==1.3.4
    # via wandb
shellingham==1.5.4
    # via typer
six==1.17.0
    # via
    #   docker-pycreds
    #   python-dateutil
smart-open==7.0.5
    # via weasel
smmap==5.0.1
    # via gitdb
spacy==3.8.2
    # via -r requirements.in
spacy-legacy==3.0.12
    # via spacy
spacy-loggers==1.0.5
    # via spacy
srsly==2.4.8
    # via
    #   confection
    #   spacy
    #   thinc
    #   weasel
st-moe-pytorch==0.1.8
    # via -r requirements.in
sympy==1.13.1
    # via torch
symusic==0.5.5
    # via miditok
thinc==8.3.2
    # via spacy
tokenizers==0.21.0
    # via
    #   miditok
    #   transformers
torch==2.5.1
    # via
    #   -r requirements.in
    #   accelerate
    #   colt5-attention
    #   local-attention
    #   st-moe-pytorch
tqdm==4.67.1
    # via
    #   -r requirements.in
    #   datasets
    #   evaluate
    #   huggingface-hub
    #   miditok
    #   spacy
    #   transformers
transformers==4.47.0
    # via -r requirements.in
triton==3.1.0
    # via torch
typer==0.15.1
    # via
    #   spacy
    #   weasel
typing-extensions==4.12.2
    # via
    #   cloudpathlib
    #   huggingface-hub
    #   multidict
    #   pydantic
    #   pydantic-core
    #   rich
    #   torch
    #   typer
    #   wandb
tzdata==2024.2
    # via pandas
urllib3==2.2.3
    # via
    #   requests
    #   sentry-sdk
wandb==0.19.0
    # via -r requirements.in
wasabi==1.1.3
    # via
    #   spacy
    #   thinc
    #   weasel
weasel==0.4.1
    # via spacy
wrapt==1.17.0
    # via smart-open
xxhash==3.5.0
    # via
    #   datasets
    #   evaluate
yarl==1.18.3
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
